performance-storage-service/requirements.txt
asgiref==3.2.7
autopep8==1.5.3
cffi==1.14.0
coverage==4.5.4
cryptography==2.9.2
Django==3.0.6
django-jenkins==0.110.0
djangorestframework==3.11.0
factory-boy==2.12.0
Faker==4.1.1
flake8==3.8.3
gunicorn==20.0.4
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
pep8==1.7.1
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
